Upgrades2 tweaks (#113)
* Enable mockgen directive for snapshots
* separate input snapshot test builder into separate template
* output snapshot builder
* rename to InputSnapshotManualBuilder
