+++ codegen/render/kube_types_renderer.go
@@ -1,10 +1,7 @@
 package render
 
 import (
-	"fmt"
 	"path/filepath"
-	"regexp"
-	"strings"
 )
 
 // renders kubernetes from templates
@@ -118,29 +115,8 @@ func (r KubeCodeRenderer) RenderKubeCode(grp Group) ([]OutFile, error) {
 	}
 
 	if grp.MockgenDirective {
-		// prepend each .go file with a go:generate directive to run mockgen on the file
-		for i, file := range files {
-			if !strings.HasSuffix(file.Path, ".go") {
-				continue
-			}
-			// only add the directive if the file contains interfaces
-			if !containsInterface(file.Content) {
-				continue
-			}
-
-			baseFile := filepath.Base(file.Path)
-			mockgenComment := fmt.Sprintf("//go:generate mockgen -source ./%v -destination mocks/%v", baseFile, baseFile)
-
-			file.Content = fmt.Sprintf("%v\n\n%v", mockgenComment, file.Content)
-			files[i] = file
-		}
+		PrependMockgenDirective(files)
 	}
 
 	return files, nil
-}
-
-var interfaceRegex = regexp.MustCompile("type .* interface")
-
-func containsInterface(content string) bool {
-	return interfaceRegex.MatchString(content)
 }
